Move InvocationMessage models to own file
